requirements.txt
streamlit
camelot-py[cv]
pandas
openpyxl
opencv-python-headless